==> 02-make-imputations/03b_vars_to_drop.R <==
drop_vars <- c("mz028b", "mz028b_org",
               "b032", "b032_org",
               "c645a", "c645a_org",
               "a006", "a006_org",
               "kd685", "kd685_org",
               "c804", "c804_org",
               "kz030", "kz030_org",
               "a525", "a525_org",
               "a551", "a551_org",
               "b650", "b650_org",
               "b663", "b663_org",
               "b665", "b665_org",
               "b667", "b667_org",
               "c483", "c483_org",
               "matsmok_tri1", "matsmok_tri1_org",
               "matsmok_tri2", "matsmok_tri2_org",
               "matsmok_tri3_c", "matsmok_tri3_c_org",
               "matsmok_tri3_e", "matsmok_tri3_e_org",
               "ACEcat_extended_0_8yrs", "ACEcat_classic_0_8yrs",
               "kz021", "kz021_org",
               "bestgest", "bestgest_org",
               "glyca", "age_bloods_7")
