==> 02-make-imputations/03c_names.R <==
names(dat)[names(dat) == "physical_abuse_0_8yrs"] <- "Physical_abuse"
names(dat)[names(dat) == "emotional_abuse_0_8yrs"] <- "Emotional_abuse"
names(dat)[names(dat) == "bullying_0_8yrs"] <- "Bullying"
names(dat)[names(dat) == "substance_household_0_8yrs"] <- "Substance_abuse"
names(dat)[names(dat) == "parent_convicted_offence_0_8yrs"] <- "Convicted_offences"
names(dat)[names(dat) == "social_class_0_8yrs"] <- "Social_class"
names(dat)[names(dat) == "neighbourhood_0_8yrs"] <- "Neighbourhood"
names(dat)[names(dat) == "physical_illness_child_0_8yrs"] <- "Physical_illness_child"
names(dat)[names(dat) == "parent_child_bond_0_8yrs"] <- "Parent_child_bond"
names(dat)[names(dat) == "ACEscore_classic_0_8yrs"] <- "ACE_score_classic"
names(dat)[names(dat) == "dw002_org"] <- "Maternal_weight"
names(dat)[names(dat) == "c666a_org"] <- "Edu_partner_mthrReport"
names(dat)[names(dat) == "pb342a_org"] <- "Edu_mother_ptnrReport"
names(dat)[names(dat) == "c520_org"] <- "Diff_afford_food"
names(dat)[names(dat) == "b370_org"] <- "EPDS_mthr_18wksG"
names(dat)[names(dat) == "pb260_org"] <- "EPDS_ptnr_18wksG"
names(dat)[names(dat) == "pb188a_org"] <- "pregCovicted_offence_ptnr"
names(dat)[names(dat) == "pb177"] <- "pregPartner_separated"
names(dat)[names(dat) == "pb098_org"] <- "pregPartned_hard_drugs"
names(dat)[names(dat) == "t5404"] <- "Antidep_use_mthr_18yrs"
names(dat)[names(dat) == "fa5404"] <- "Antidep_use_ptnr_18yrs"
names(dat)[names(dat) == "ypa5005_dup"] <- "PtnrChild_phys_force_18yrs"
names(dat)[names(dat) == "ypa5009_dup"] <- "PtnrChild_kiss_touch_18yrs"
names(dat)[names(dat) == "ypa5013_dup"] <- "PtnrChild_pressure_sex_18yrs"
names(dat)[names(dat) == "ypa5017_dup"] <- "PtnrChild_feel_scared_18yrs"
names(dat)[names(dat) == "t5510_dup_org"] <- "AUDIT_ptnr_18yrs"
names(dat)[names(dat) == "il6"] <- "IL6"
names(dat)[names(dat) == "ui_14"] <- "UI"
names(dat)[names(dat) == "bedwet_14"] <- "Bedwetting"
names(dat)[names(dat) == "nocturia_14"] <- "Nocturia"
names(dat)[names(dat) == "void_postpone_14"] <- "Voiding_postponement"
names(dat)[names(dat) == "bmi_8"] <- "BMI"
names(dat)[names(dat) == "sdq_8"] <- "SDQ"
names(dat)[names(dat) == "marital"] <- "Marital_status"
names(dat)[names(dat) == "age_bloods_7"] <- "Age_at_blood_draw_7"
names(dat)[names(dat) == "sexual_abuse_0_8yrs"] <- "Sexual_abuse"
names(dat)[names(dat) == "emotional_neglect_0_8yrs"] <- "Emotional_neglect"
names(dat)[names(dat) == "violence_between_parents_0_8yrs"] <- "Violence_parents"
names(dat)[names(dat) == "mental_health_problems_or_suicide_0_8yrs"] <- "Mental_hlth_suicide"
names(dat)[names(dat) == "parental_separation_0_8yrs"] <- "Parental_separation"
names(dat)[names(dat) == "financial_difficulties_0_8yrs"] <- "Financial_difficulties"
names(dat)[names(dat) == "social_support_parent_0_8yrs"] <- "Social_support_parents"
names(dat)[names(dat) == "physical_illness_parent_0_8yrs"] <- "Physical_illness_parents"
names(dat)[names(dat) == "ACEscore_extended_0_8yrs"] <- "ACE_score_extended"
names(dat)[names(dat) == "dw042_org"] <- "Maternal_BMI"
names(dat)[names(dat) == "pb325a_org"] <- "Edu_ptnr_self"
names(dat)[names(dat) == "c472_org"] <- "Homeless_mthr_preg"
names(dat)[names(dat) == "c522_org"] <- "Diff_afford_heating"
names(dat)[names(dat) == "c600_org"] <- "EPDS_mthr_32wksG"
names(dat)[names(dat) == "a600_org"] <- "neighbourhood_mthr_preg"
names(dat)[names(dat) == "b587"] <- "Divorced_mthr_preg"
names(dat)[names(dat) == "sc_household_18wgest_org"] <- "Social_class_18wks"
names(dat)[names(dat) == "t3336"] <- "Ptnr_emo_cruel_18yrs"
names(dat)[names(dat) == "t3255_org"] <- "EDPS_mthr_18yrs"
names(dat)[names(dat) == "t3316"] <- "Mthr_separated_18yrs"
names(dat)[names(dat) == "ypa5007_dup"] <- "PtnrChild_sev_phys_force_18yrs"
names(dat)[names(dat) == "ypa5011_dup"] <- "PtnrChild_sev_kiss_touch_18yrs"
names(dat)[names(dat) == "ypa5015_dup"] <- "PtnrChild_physforce_sex_18yrs"
names(dat)[names(dat) == "fa5510_dup_org"] <- "AUDIT_mthr_18yrs"
names(dat)[names(dat) == "crp"] <- "CRP"
names(dat)[names(dat) == "glyca"] <- "GlycA"
names(dat)[names(dat) == "daywet_14"] <- "Daywetting"
names(dat)[names(dat) == "freq_urine_14"] <- "Frequency"
names(dat)[names(dat) == "urgency_14"] <- "Urgency"
names(dat)[names(dat) == "void_volume_14"] <- "Voiding_volume"
names(dat)[names(dat) == "constipation_8"] <- "Constipation"
names(dat)[names(dat) == "sex"] <- "Sex"
names(dat)[names(dat) == "parity"] <- "Parity"
names(dat)[names(dat) == "housetenure"] <- "House_tenure"
names(dat)[names(dat) == "ethnicity"] <- "Ethnicity"
names(dat)[names(dat) == "gest_age"] <- "Gestational_age"
names(dat)[names(dat) == "crowding"] <- "Crowding"
names(dat)[names(dat) == "age_bloods_9"] <- "Age_blood_draw_9"
names(dat)[names(dat) == "birthweight"] <- "Birthweight"
names(dat)[names(dat) == "b593"] <- "Homeless_mthr"

names(dat_girls)[names(dat_girls) == "physical_abuse_0_8yrs"] <- "Physical_abuse"
names(dat_girls)[names(dat_girls) == "emotional_abuse_0_8yrs"] <- "Emotional_abuse"
names(dat_girls)[names(dat_girls) == "bullying_0_8yrs"] <- "Bullying"
names(dat_girls)[names(dat_girls) == "substance_household_0_8yrs"] <- "Substance_abuse"
names(dat_girls)[names(dat_girls) == "parent_convicted_offence_0_8yrs"] <- "Convicted_offences"
names(dat_girls)[names(dat_girls) == "social_class_0_8yrs"] <- "Social_class"
names(dat_girls)[names(dat_girls) == "neighbourhood_0_8yrs"] <- "Neighbourhood"
names(dat_girls)[names(dat_girls) == "physical_illness_child_0_8yrs"] <- "Physical_illness_child"
names(dat_girls)[names(dat_girls) == "parent_child_bond_0_8yrs"] <- "Parent_child_bond"
names(dat_girls)[names(dat_girls) == "ACEscore_classic_0_8yrs"] <- "ACE_score_classic"
names(dat_girls)[names(dat_girls) == "dw002_org"] <- "Maternal_weight"
names(dat_girls)[names(dat_girls) == "c666a_org"] <- "Edu_partner_mthrReport"
names(dat_girls)[names(dat_girls) == "pb342a_org"] <- "Edu_mother_ptnrReport"
names(dat_girls)[names(dat_girls) == "c520_org"] <- "Diff_afford_food"
names(dat_girls)[names(dat_girls) == "b370_org"] <- "EPDS_mthr_18wksG"
names(dat_girls)[names(dat_girls) == "pb260_org"] <- "EPDS_ptnr_18wksG"
names(dat_girls)[names(dat_girls) == "pb188a_org"] <- "pregCovicted_offence_ptnr"
names(dat_girls)[names(dat_girls) == "pb177"] <- "pregPartner_separated"
names(dat_girls)[names(dat_girls) == "pb098_org"] <- "pregPartned_hard_drugs"
names(dat_girls)[names(dat_girls) == "t5404"] <- "Antidep_use_mthr_18yrs"
names(dat_girls)[names(dat_girls) == "fa5404"] <- "Antidep_use_ptnr_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5005_dup"] <- "PtnrChild_phys_force_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5009_dup"] <- "PtnrChild_kiss_touch_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5013_dup"] <- "PtnrChild_pressure_sex_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5017_dup"] <- "PtnrChild_feel_scared_18yrs"
names(dat_girls)[names(dat_girls) == "t5510_dup_org"] <- "AUDIT_ptnr_18yrs"
names(dat_girls)[names(dat_girls) == "il6"] <- "IL6"
names(dat_girls)[names(dat_girls) == "ui_14"] <- "UI"
names(dat_girls)[names(dat_girls) == "bedwet_14"] <- "Bedwetting"
names(dat_girls)[names(dat_girls) == "nocturia_14"] <- "Nocturia"
names(dat_girls)[names(dat_girls) == "void_postpone_14"] <- "Voiding_postponement"
names(dat_girls)[names(dat_girls) == "bmi_8"] <- "BMI"
names(dat_girls)[names(dat_girls) == "sdq_8"] <- "SDQ"
names(dat_girls)[names(dat_girls) == "marital"] <- "Marital_status"
names(dat_girls)[names(dat_girls) == "age_bloods_7"] <- "Age_at_blood_draw_7"
names(dat_girls)[names(dat_girls) == "sexual_abuse_0_8yrs"] <- "Sexual_abuse"
names(dat_girls)[names(dat_girls) == "emotional_neglect_0_8yrs"] <- "Emotional_neglect"
names(dat_girls)[names(dat_girls) == "violence_between_parents_0_8yrs"] <- "Violence_parents"
names(dat_girls)[names(dat_girls) == "mental_health_problems_or_suicide_0_8yrs"] <- "Mental_hlth_suicide"
names(dat_girls)[names(dat_girls) == "parental_separation_0_8yrs"] <- "Parental_separation"
names(dat_girls)[names(dat_girls) == "financial_difficulties_0_8yrs"] <- "Financial_difficulties"
names(dat_girls)[names(dat_girls) == "social_support_parent_0_8yrs"] <- "Social_support_parents"
names(dat_girls)[names(dat_girls) == "physical_illness_parent_0_8yrs"] <- "Physical_illness_parents"
names(dat_girls)[names(dat_girls) == "ACEscore_extended_0_8yrs"] <- "ACE_score_extended"
names(dat_girls)[names(dat_girls) == "dw042_org"] <- "Maternal_BMI"
names(dat_girls)[names(dat_girls) == "pb325a_org"] <- "Edu_ptnr_self"
names(dat_girls)[names(dat_girls) == "c472_org"] <- "Homeless_mthr_preg"
names(dat_girls)[names(dat_girls) == "c522_org"] <- "Diff_afford_heating"
names(dat_girls)[names(dat_girls) == "c600_org"] <- "EPDS_mthr_32wksG"
names(dat_girls)[names(dat_girls) == "a600_org"] <- "neighbourhood_mthr_preg"
names(dat_girls)[names(dat_girls) == "b587"] <- "Divorced_mthr_preg"
names(dat_girls)[names(dat_girls) == "sc_household_18wgest_org"] <- "Social_class_18wks"
names(dat_girls)[names(dat_girls) == "t3336"] <- "Ptnr_emo_cruel_18yrs"
names(dat_girls)[names(dat_girls) == "t3255_org"] <- "EDPS_mthr_18yrs"
names(dat_girls)[names(dat_girls) == "t3316"] <- "Mthr_separated_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5007_dup"] <- "PtnrChild_sev_phys_force_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5011_dup"] <- "PtnrChild_sev_kiss_touch_18yrs"
names(dat_girls)[names(dat_girls) == "ypa5015_dup"] <- "PtnrChild_physforce_sex_18yrs"
names(dat_girls)[names(dat_girls) == "fa5510_dup_org"] <- "AUDIT_mthr_18yrs"
names(dat_girls)[names(dat_girls) == "crp"] <- "CRP"
names(dat_girls)[names(dat_girls) == "glyca"] <- "GlycA"
names(dat_girls)[names(dat_girls) == "daywet_14"] <- "Daywetting"
names(dat_girls)[names(dat_girls) == "freq_urine_14"] <- "Frequency"
names(dat_girls)[names(dat_girls) == "urgency_14"] <- "Urgency"
names(dat_girls)[names(dat_girls) == "void_volume_14"] <- "Voiding_volume"
names(dat_girls)[names(dat_girls) == "constipation_8"] <- "Constipation"
names(dat_girls)[names(dat_girls) == "sex"] <- "Sex"
names(dat_girls)[names(dat_girls) == "parity"] <- "Parity"
names(dat_girls)[names(dat_girls) == "housetenure"] <- "House_tenure"
names(dat_girls)[names(dat_girls) == "ethnicity"] <- "Ethnicity"
names(dat_girls)[names(dat_girls) == "gest_age"] <- "Gestational_age"
names(dat_girls)[names(dat_girls) == "crowding"] <- "Crowding"
names(dat_girls)[names(dat_girls) == "age_bloods_9"] <- "Age_blood_draw_9"
names(dat_girls)[names(dat_girls) == "birthweight"] <- "Birthweight"
names(dat_girls)[names(dat_girls) == "b593"] <- "Homeless_mthr"

names(dat_boys)[names(dat_boys) == "physical_abuse_0_8yrs"] <- "Physical_abuse"
names(dat_boys)[names(dat_boys) == "emotional_abuse_0_8yrs"] <- "Emotional_abuse"
names(dat_boys)[names(dat_boys) == "bullying_0_8yrs"] <- "Bullying"
names(dat_boys)[names(dat_boys) == "substance_household_0_8yrs"] <- "Substance_abuse"
names(dat_boys)[names(dat_boys) == "parent_convicted_offence_0_8yrs"] <- "Convicted_offences"
names(dat_boys)[names(dat_boys) == "social_class_0_8yrs"] <- "Social_class"
names(dat_boys)[names(dat_boys) == "neighbourhood_0_8yrs"] <- "Neighbourhood"
names(dat_boys)[names(dat_boys) == "physical_illness_child_0_8yrs"] <- "Physical_illness_child"
names(dat_boys)[names(dat_boys) == "parent_child_bond_0_8yrs"] <- "Parent_child_bond"
names(dat_boys)[names(dat_boys) == "ACEscore_classic_0_8yrs"] <- "ACE_score_classic"
names(dat_boys)[names(dat_boys) == "dw002_org"] <- "Maternal_weight"
names(dat_boys)[names(dat_boys) == "c666a_org"] <- "Edu_partner_mthrReport"
names(dat_boys)[names(dat_boys) == "pb342a_org"] <- "Edu_mother_ptnrReport"
names(dat_boys)[names(dat_boys) == "c520_org"] <- "Diff_afford_food"
names(dat_boys)[names(dat_boys) == "b370_org"] <- "EPDS_mthr_18wksG"
names(dat_boys)[names(dat_boys) == "pb260_org"] <- "EPDS_ptnr_18wksG"
names(dat_boys)[names(dat_boys) == "pb188a_org"] <- "pregCovicted_offence_ptnr"
names(dat_boys)[names(dat_boys) == "pb177"] <- "pregPartner_separated"
names(dat_boys)[names(dat_boys) == "pb098_org"] <- "pregPartned_hard_drugs"
names(dat_boys)[names(dat_boys) == "t5404"] <- "Antidep_use_mthr_18yrs"
names(dat_boys)[names(dat_boys) == "fa5404"] <- "Antidep_use_ptnr_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5005_dup"] <- "PtnrChild_phys_force_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5009_dup"] <- "PtnrChild_kiss_touch_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5013_dup"] <- "PtnrChild_pressure_sex_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5017_dup"] <- "PtnrChild_feel_scared_18yrs"
names(dat_boys)[names(dat_boys) == "t5510_dup_org"] <- "AUDIT_ptnr_18yrs"
names(dat_boys)[names(dat_boys) == "il6"] <- "IL6"
names(dat_boys)[names(dat_boys) == "ui_14"] <- "UI"
names(dat_boys)[names(dat_boys) == "bedwet_14"] <- "Bedwetting"
names(dat_boys)[names(dat_boys) == "nocturia_14"] <- "Nocturia"
names(dat_boys)[names(dat_boys) == "void_postpone_14"] <- "Voiding_postponement"
names(dat_boys)[names(dat_boys) == "bmi_8"] <- "BMI"
names(dat_boys)[names(dat_boys) == "sdq_8"] <- "SDQ"
names(dat_boys)[names(dat_boys) == "marital"] <- "Marital_status"
names(dat_boys)[names(dat_boys) == "age_bloods_7"] <- "Age_at_blood_draw_7"
names(dat_boys)[names(dat_boys) == "sexual_abuse_0_8yrs"] <- "Sexual_abuse"
names(dat_boys)[names(dat_boys) == "emotional_neglect_0_8yrs"] <- "Emotional_neglect"
names(dat_boys)[names(dat_boys) == "violence_between_parents_0_8yrs"] <- "Violence_parents"
names(dat_boys)[names(dat_boys) == "mental_health_problems_or_suicide_0_8yrs"] <- "Mental_hlth_suicide"
names(dat_boys)[names(dat_boys) == "parental_separation_0_8yrs"] <- "Parental_separation"
names(dat_boys)[names(dat_boys) == "financial_difficulties_0_8yrs"] <- "Financial_difficulties"
names(dat_boys)[names(dat_boys) == "social_support_parent_0_8yrs"] <- "Social_support_parents"
names(dat_boys)[names(dat_boys) == "physical_illness_parent_0_8yrs"] <- "Physical_illness_parents"
names(dat_boys)[names(dat_boys) == "ACEscore_extended_0_8yrs"] <- "ACE_score_extended"
names(dat_boys)[names(dat_boys) == "dw042_org"] <- "Maternal_BMI"
names(dat_boys)[names(dat_boys) == "pb325a_org"] <- "Edu_ptnr_self"
names(dat_boys)[names(dat_boys) == "c472_org"] <- "Homeless_mthr_preg"
names(dat_boys)[names(dat_boys) == "c522_org"] <- "Diff_afford_heating"
names(dat_boys)[names(dat_boys) == "c600_org"] <- "EPDS_mthr_32wksG"
names(dat_boys)[names(dat_boys) == "a600_org"] <- "neighbourhood_mthr_preg"
names(dat_boys)[names(dat_boys) == "b587"] <- "Divorced_mthr_preg"
names(dat_boys)[names(dat_boys) == "sc_household_18wgest_org"] <- "Social_class_18wks"
names(dat_boys)[names(dat_boys) == "t3336"] <- "Ptnr_emo_cruel_18yrs"
names(dat_boys)[names(dat_boys) == "t3255_org"] <- "EDPS_mthr_18yrs"
names(dat_boys)[names(dat_boys) == "t3316"] <- "Mthr_separated_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5007_dup"] <- "PtnrChild_sev_phys_force_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5011_dup"] <- "PtnrChild_sev_kiss_touch_18yrs"
names(dat_boys)[names(dat_boys) == "ypa5015_dup"] <- "PtnrChild_physforce_sex_18yrs"
names(dat_boys)[names(dat_boys) == "fa5510_dup_org"] <- "AUDIT_mthr_18yrs"
names(dat_boys)[names(dat_boys) == "crp"] <- "CRP"
names(dat_boys)[names(dat_boys) == "glyca"] <- "GlycA"
names(dat_boys)[names(dat_boys) == "daywet_14"] <- "Daywetting"
names(dat_boys)[names(dat_boys) == "freq_urine_14"] <- "Frequency"
names(dat_boys)[names(dat_boys) == "urgency_14"] <- "Urgency"
names(dat_boys)[names(dat_boys) == "void_volume_14"] <- "Voiding_volume"
names(dat_boys)[names(dat_boys) == "constipation_8"] <- "Constipation"
names(dat_boys)[names(dat_boys) == "sex"] <- "Sex"
names(dat_boys)[names(dat_boys) == "parity"] <- "Parity"
names(dat_boys)[names(dat_boys) == "housetenure"] <- "House_tenure"
names(dat_boys)[names(dat_boys) == "ethnicity"] <- "Ethnicity"
names(dat_boys)[names(dat_boys) == "gest_age"] <- "Gestational_age"
names(dat_boys)[names(dat_boys) == "crowding"] <- "Crowding"
names(dat_boys)[names(dat_boys) == "age_bloods_9"] <- "Age_blood_draw_9"
names(dat_boys)[names(dat_boys) == "birthweight"] <- "Birthweight"
names(dat_boys)[names(dat_boys) == "b593"] <- "Homeless_mthr"
